Started Chord System, Unified design
Added a bunch of new sound assets to enable chord functionality.
Added new materials for the note blocks to make it easy to tell them apart and to give them a theme.
Tried to import the UE4 third person example animations/model and it broke.
Added two new levels and fixed UI issues with not opening the next level.
Removed test level (this might have been a bad idea).
Created the sound cues needed to create chords and made a system for encoding them into the sound map (not sure if it fully works yet though).
Started work on making the submit platforms allowing more than one note block to be added (not functional yet).
Changed the way the player proceeded to the next level - instead of having it be on a seperate screen, the next level option now displays in the current level.